requirements.txt
aiohttp
sentry-sdk
py-cord
numpy
pynacl
configparser
python-dateutil
